Refactor embedded resources
