+++ Context/Context/Resources/EmbeddedResourceView.swift
@@ -3,7 +3,6 @@
 import ContextCore
 import QuickLook
 import SwiftUI
-import UniformTypeIdentifiers
 
 struct EmbeddedResourceView: View {
   let resources: [EmbeddedResource]
@@ -53,7 +52,7 @@ struct EmbeddedResourceView: View {
         if let resource = selectedResource {
           // Toolbar
           HStack(spacing: 12) {
-            Text(resourceTitle(for: resource))
+            Text(ResourceOperations.resourceTitle(for: resource))
               .font(.headline)
               .foregroundColor(.primary)
 
@@ -62,7 +61,7 @@ struct EmbeddedResourceView: View {
             // Copy button - behavior changes based on view mode
             CopyButton {
               if viewMode == .preview {
-                copyPreviewContent(resource)
+                ResourceOperations.copyToClipboard(resource)
               } else {
                 copyRawJSONToClipboard()
               }
@@ -72,7 +71,7 @@ struct EmbeddedResourceView: View {
 
             // Save button
             Button(action: {
-              saveResource(resource)
+              ResourceOperations.saveResource(resource)
             }) {
               Image(systemName: "square.and.arrow.down")
                 .font(.system(size: 14))
@@ -144,8 +143,9 @@ struct EmbeddedResourceView: View {
           } else {
             // Raw view
             if let rawJSON = rawJSON,
-               let jsonData = rawJSON.data(using: .utf8),
-               let jsonValue = try? JSONDecoder().decode(JSONValue.self, from: jsonData) {
+              let jsonData = rawJSON.data(using: .utf8),
+              let jsonValue = try? JSONDecoder().decode(JSONValue.self, from: jsonData)
+            {
               JSONRawView(jsonValue: jsonValue, searchText: "", isSearchActive: false)
                 .transition(.opacity)
             } else {
@@ -219,227 +219,9 @@ struct EmbeddedResourceView: View {
     // Clear quickLook URL
     quickLookURL = nil
 
-    // Create share URL for resources
-    switch resource {
-    case .text(let textContent):
-      createShareURLForText(textContent)
-    case .blob(let blobContent):
-      createShareURL(for: blobContent)
-    }
+    shareURL = ResourceOperations.createShareURL(for: resource)
   }
 
-  private func createShareURLForText(_ textContent: TextResourceContents) {
-    // Generate filename from URI
-    var fileName = URL(string: textContent.uri)?.lastPathComponent ?? "resource"
-
-    // Ensure filename is not empty or just a path separator
-    if fileName.isEmpty || fileName == "/" {
-      fileName = "resource"
-    }
-
-    // Add appropriate extension based on MIME type
-    let fileExtension = FileExtensionHelper.fileExtension(for: textContent.mimeType)
-    let fullFileName = fileName.contains(".") ? fileName : "\(fileName)\(fileExtension)"
-
-    // Ensure we have a valid filename
-    let safeFileName = fullFileName.isEmpty ? "resource.txt" : fullFileName
-
-    // Create temporary file
-    let tempDir = FileManager.default.temporaryDirectory
-    let tempURL = tempDir.appendingPathComponent(safeFileName)
-
-    do {
-      try textContent.text.write(to: tempURL, atomically: true, encoding: .utf8)
-      shareURL = tempURL
-    } catch {
-      print("Failed to create share URL for text: \(error)")
-    }
-  }
-
-  private func createShareURL(for blobContent: BlobResourceContents) {
-    // Generate filename from URI
-    var fileName = URL(string: blobContent.uri)?.lastPathComponent ?? "resource"
-
-    // Ensure filename is not empty or just a path separator
-    if fileName.isEmpty || fileName == "/" {
-      fileName = "resource"
-    }
-
-    // Add appropriate extension based on MIME type
-    let fileExtension = FileExtensionHelper.fileExtension(for: blobContent.mimeType)
-    let fullFileName = fileName.contains(".") ? fileName : "\(fileName)\(fileExtension)"
-
-    // Ensure we have a valid filename
-    let safeFileName = fullFileName.isEmpty ? "resource.bin" : fullFileName
-
-    // Create temporary file
-    let tempDir = FileManager.default.temporaryDirectory
-    let tempURL = tempDir.appendingPathComponent(safeFileName)
-
-    do {
-      try blobContent.blob.write(to: tempURL)
-      shareURL = tempURL
-    } catch {
-      print("Failed to create share URL: \(error)")
-    }
-  }
-
-  private func fileExtension(for mimeType: String?) -> String {
-    FileExtensionHelper.fileExtension(for: mimeType)
-  }
-
-  private func resourceTitle(for resource: EmbeddedResource) -> String {
-    switch resource {
-    case .text(let content):
-      return URL(string: content.uri)?.lastPathComponent ?? content.uri
-    case .blob(let content):
-      return URL(string: content.uri)?.lastPathComponent ?? content.uri
-    }
-  }
-
-  private func saveResource(_ resource: EmbeddedResource) {
-    let savePanel = NSSavePanel()
-
-    // Configure save panel based on resource type
-    switch resource {
-    case .text(let textContent):
-      let fileName = URL(string: textContent.uri)?.lastPathComponent ?? "resource"
-      let fileExtension = FileExtensionHelper.fileExtension(for: textContent.mimeType)
-      savePanel.nameFieldStringValue =
-        fileName.contains(".") ? fileName : "\(fileName)\(fileExtension)"
-
-      if let mimeType = textContent.mimeType {
-        switch mimeType {
-        case "text/html":
-          savePanel.allowedContentTypes = [.html]
-        case "application/json":
-          savePanel.allowedContentTypes = [.json]
-        case "application/xml", "text/xml":
-          savePanel.allowedContentTypes = [.xml]
-        case "text/css":
-          savePanel.allowedContentTypes = [.css]
-        case "text/javascript", "application/javascript":
-          savePanel.allowedContentTypes = [.javaScript]
-        case "text/markdown":
-          savePanel.allowedContentTypes = [UTType(filenameExtension: "md") ?? .plainText]
-        case "text/csv":
-          savePanel.allowedContentTypes = [.commaSeparatedText]
-        case "application/x-yaml", "text/yaml":
-          savePanel.allowedContentTypes = [.yaml]
-        case "text/x-python", "application/x-python-code":
-          savePanel.allowedContentTypes = [.pythonScript]
-        case "text/x-swift":
-          savePanel.allowedContentTypes = [.swiftSource]
-        case "text/x-c":
-          savePanel.allowedContentTypes = [.cSource]
-        case "text/x-c++":
-          savePanel.allowedContentTypes = [.cPlusPlusSource]
-        case "text/x-ruby":
-          savePanel.allowedContentTypes = [.rubyScript]
-        case "text/x-php":
-          savePanel.allowedContentTypes = [.phpScript]
-        default:
-          savePanel.allowedContentTypes = [.plainText]
-        }
-      } else {
-        savePanel.allowedContentTypes = [.plainText]
-      }
-
-    case .blob(let blobContent):
-      let fileName = URL(string: blobContent.uri)?.lastPathComponent ?? "resource"
-      let fileExtension = FileExtensionHelper.fileExtension(for: blobContent.mimeType)
-      savePanel.nameFieldStringValue =
-        fileName.contains(".") ? fileName : "\(fileName)\(fileExtension)"
-
-      // Set allowed content types based on MIME type
-      if let mimeType = blobContent.mimeType {
-        switch mimeType {
-        // Images
-        case "image/png":
-          savePanel.allowedContentTypes = [.png]
-        case "image/jpeg", "image/jpg":
-          savePanel.allowedContentTypes = [.jpeg]
-        case "image/gif":
-          savePanel.allowedContentTypes = [.gif]
-        case "image/svg+xml":
-          savePanel.allowedContentTypes = [.svg]
-        case "image/webp":
-          savePanel.allowedContentTypes = [.webP]
-        case "image/bmp":
-          savePanel.allowedContentTypes = [.bmp]
-        case "image/tiff":
-          savePanel.allowedContentTypes = [.tiff]
-        case "image/x-icon":
-          savePanel.allowedContentTypes = [.ico]
-        // Videos
-        case "video/mp4":
-          savePanel.allowedContentTypes = [.mpeg4Movie]
-        case "video/quicktime":
-          savePanel.allowedContentTypes = [.quickTimeMovie]
-        case "video/x-msvideo":
-          savePanel.allowedContentTypes = [.avi]
-        case "video/mpeg":
-          savePanel.allowedContentTypes = [.mpeg]
-        // Audio
-        case "audio/mpeg", "audio/mp3":
-          savePanel.allowedContentTypes = [.mp3]
-        case "audio/wav", "audio/x-wav":
-          savePanel.allowedContentTypes = [.wav]
-        case "audio/aac":
-          savePanel.allowedContentTypes = [UTType(filenameExtension: "aac") ?? .audio]
-        case "audio/flac":
-          savePanel.allowedContentTypes = [UTType(filenameExtension: "flac") ?? .audio]
-        case "audio/midi", "audio/x-midi":
-          savePanel.allowedContentTypes = [.midi]
-        // Documents
-        case "application/pdf":
-          savePanel.allowedContentTypes = [.pdf]
-        case "application/zip":
-          savePanel.allowedContentTypes = [.zip]
-        default:
-          // Try to create UTType from file extension
-          let ext = FileExtensionHelper.fileExtension(for: mimeType).trimmingCharacters(
-            in: CharacterSet(charactersIn: "."))
-          if !ext.isEmpty, let utType = UTType(filenameExtension: ext) {
-            savePanel.allowedContentTypes = [utType]
-          } else {
-            savePanel.allowedContentTypes = []
-          }
-        }
-      }
-    }
-
-    savePanel.canCreateDirectories = true
-    savePanel.showsTagField = false
-    savePanel.isExtensionHidden = false
-
-    savePanel.begin { response in
-      if response == .OK, let url = savePanel.url {
-        do {
-          switch resource {
-          case .text(let textContent):
-            try textContent.text.write(to: url, atomically: true, encoding: .utf8)
-          case .blob(let blobContent):
-            try blobContent.blob.write(to: url)
-          }
-        } catch {
-          print("Failed to save resource: \(error)")
-        }
-      }
-    }
-  }
-
-  private func copyPreviewContent(_ resource: EmbeddedResource) {
-    switch resource {
-    case .text(let textContent):
-      copyTextToClipboard(textContent.text)
-    case .blob(let blobContent):
-      if let mimeType = blobContent.mimeType, mimeType.starts(with: "image/") {
-        copyImageToClipboard(blobContent.blob)
-      }
-    }
-  }
-  
   private func shouldDisableCopyButton(for resource: EmbeddedResource) -> Bool {
     switch resource {
     case .text:
@@ -449,24 +231,12 @@ struct EmbeddedResourceView: View {
       return !(blobContent.mimeType?.starts(with: "image/") ?? false)
     }
   }
-  
+
   private func copyRawJSONToClipboard() {
     let jsonString = rawJSON ?? "null"
     let unescapedJson = jsonString.replacingOccurrences(of: "\\/", with: "/")
     NSPasteboard.general.clearContents()
     NSPasteboard.general.setString(unescapedJson, forType: .string)
-  }
-
-  private func copyTextToClipboard(_ text: String) {
-    NSPasteboard.general.clearContents()
-    NSPasteboard.general.setString(text, forType: .string)
-  }
-
-  private func copyImageToClipboard(_ data: Data) {
-    if let image = NSImage(data: data) {
-      NSPasteboard.general.clearContents()
-      NSPasteboard.general.writeObjects([image])
-    }
   }
 }
